[WV-170] move methods from update_recommended_politicians_view into controllers_recommendation
